Do not crash when attempting to delete a system message.
Apparently if we tried to delete a system message, an exception was thrown, and the rest
of the channel or thread was not processed. There is a function to check for whether a
message is a system message in the pycord API. I added that check, but it returns false
even on system messages (??). So I added a try-except clause to catch the error
so that the error is not fatal and the bot can proceed with processing the channel/thread.