Check message timestamp rather than system time
Check that the timestamp of the message, rather than the system clock, is on the first minute of the day.

Split the original date extraction helper to one that extracts the datetime in the proper timezone, and one that isolates the date from that. This is needed because the isolated date does not carry the time.