Allow some wiggle in accepted happy midnights
First check with a regex if this might be a happy midnight expression. Then compute the edit distance to two known expressions and if one of them is close enough, allow it.

This means that small amounts of self expression (e.g. "hyvää keksiyötä"), a couple of small typos or a missing space should not keep anyone from catching the trophy. More than three characters difference means the message will not be considered a midnight phrase.

The regex part isolates the assumed happy midnight string to allow other stuff, such as emojis, to be added before and after the expression without being counted towards the edit distance.

This check requires more cycles than a simple substring match or a simple regex check, so this is better put behind a time barrier, i.e. only do this close to midnight.